Fix beatsin8() phase_offset domain bug (DNA showing one strand, not two)
wled_compat.h's beatsin8() declared phase_offset as uint16_t and added it
in the 16-bit beat domain before truncating to 8 bits for sin8(). Real
WLED's beatsin8_t() (wled00/util.cpp) adds phase_offset in the 8-bit
domain, after truncation - so a phase_offset of 128 there means exactly
half a cycle (180 degrees), which is how effects like DNA put two waves
opposite each other to form a double helix. Under the old (buggy) 16-bit
math, the same 128 was only ~0.2% of a cycle - invisible, which is why DNA
(effect id 152) was reported showing a single line instead of two strands.

Since this is a shared helper, the bug affected every effect calling
beatsin8() with a nonzero phase_offset, not just DNA - fixed centrally
rather than patched per-effect. Verified via a real-WLED source citation
(util.cpp:567's beatsin8_t(), phase_offset typed uint8_t there too).

+++ src/effects/wled_compat.h
@@ -33,8 +33,20 @@ int16_t cos16(uint16_t theta);  // roughly -32767..32767
 // strip.now; here it's an explicit `now_ms` parameter (this firmware's
 // equivalent, passed through from render()'s now_ms same as everywhere
 // else).
+//
+// `phase_offset` is uint8_t here, matching real WLED's beatsin8_t()
+// exactly (wled00/util.cpp: `sin8_t(beat8(bpm, timebase) + phase_offset)`)
+// - it's added in the same 8-bit domain sin8() operates in, so 128 means
+// exactly half a cycle (used constantly to put two waves 180 degrees
+// apart, e.g. the two strands of the DNA effect). A wider type here would
+// silently shrink that to a fraction of a percent of a cycle instead - a
+// real bug this had until a report of DNA showing one line instead of two
+// (effect id 152) traced it here. Callers passing an expression like
+// `x*10` that overflows 255 get real WLED's own wraparound behavior for
+// free via the implicit truncation, which several effects rely on as a
+// deliberate repeating pattern - not something to "fix" further.
 uint8_t beatsin8(uint32_t now_ms, uint8_t bpm, uint8_t low = 0, uint8_t high = 255,
-                  uint16_t phase_offset = 0);
+                  uint8_t phase_offset = 0);
 uint16_t beatsin16(uint32_t now_ms, uint8_t bpm, uint16_t low = 0, uint16_t high = 65535,
                     uint16_t phase_offset = 0);
